Gemini/requirements.txt
# Run the following command to install the dependencies:
# pip install -r requirements.txt

google-generativeai
pymupdf
python-dotenv
Flask
